refactor(providers): 统一命名 — ProviderAdapter → Provider, ProviderRegistry → ProviderManager
- ProviderAdapter → Provider（降低抽象性）
- OpenAIAdapter → OpenAIProvider
- OpenRouterAdapter → OpenRouterProvider
- ProviderRegistry → ProviderManager（与 SessionManager 等命名一致）
- 文件同步重命名: registry→manager, openai_adapter→openai_provider
- bay-project.md 文档同步更新
